Fix many calls to ctype functions
All ctype functions like `isspace()`, `tolower()` and alike expect a
character as an `unsigned char` or EOF, anything else has undefined
behavior.

This commit should fix all calls to the various ctype functions either:

* adding `unsigned char` cast where appropriate;
* removing useless `int` casts that are more confusing than useful.

Some instances required slightly more involved changes though, but the
fix itself is the same.

To find the calls to check, I used an equivalent to the following:

```
git grep -P 'is(alnum|alpha|cntrl|x?digit|graph|lower|print|punct|'\
'space|upper|ascii|blank)|to(low|up)er'
```

...and then proceeded to manually check each one of them to see if the
call, or call chain, was using appropriate values and types.

+++ parsers/asp.c
@@ -51,7 +51,7 @@ static void findAspTags (void)
 		while (*cp != '\0')
 		{
 			/* jump over whitespace */
-			while (isspace ((int)*cp))
+			while (isspace (*cp))
 				cp++;
 
 			/* jump over strings */
@@ -70,9 +70,9 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "end", (size_t) 3)== 0)
 			{
 				cp += 3;
-				if (isspace ((int)*cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int)*cp))
+					while (isspace (*cp))
 						++cp;
 
 					if (strncasecmp ((const char*) cp, "function", (size_t) 8) == 0)
@@ -93,9 +93,9 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "exit", (size_t) 4)==0)
 			{
 				cp += 4;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
 
 					if (strncasecmp ((const char*) cp, "function", (size_t) 8) == 0)
@@ -116,16 +116,16 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "public", (size_t) 6) == 0)
 			{
 				cp += 6;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
 					if (strncasecmp ((const char*) cp, "function", (size_t) 8) == 0)
 					{
 						cp+=8;
-					    while (isspace ((int) *cp))
+					    while (isspace (*cp))
 						    ++cp;
-					    while (isalnum ((int) *cp)  ||  *cp == '_')
+					    while (isalnum (*cp)  ||  *cp == '_')
 					    {
 						    vStringPut (name, *cp);
 						    ++cp;
@@ -136,9 +136,9 @@ static void findAspTags (void)
 					else if (strncasecmp ((const char*) cp, "sub", (size_t) 3) == 0)
 					{
 						cp+=3;
-					    while (isspace ((int) *cp))
+					    while (isspace (*cp))
 						    ++cp;
-					    while (isalnum ((int) *cp)  ||  *cp == '_')
+					    while (isalnum (*cp)  ||  *cp == '_')
 					    {
 						    vStringPut (name, *cp);
 						    ++cp;
@@ -147,7 +147,7 @@ static void findAspTags (void)
 					    vStringClear (name);
 					}
 					else {
-					    while (isalnum ((int) *cp)  ||  *cp == '_')
+					    while (isalnum (*cp)  ||  *cp == '_')
 					    {
 						    vStringPut (name, *cp);
 						    ++cp;
@@ -160,16 +160,16 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "private", (size_t) 7) == 0)
 			{
 				cp += 7;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
 					if (strncasecmp ((const char*) cp, "function", (size_t) 8) == 0)
 					{
 						cp+=8;
-					    while (isspace ((int) *cp))
+					    while (isspace (*cp))
 						    ++cp;
-					    while (isalnum ((int) *cp)  ||  *cp == '_')
+					    while (isalnum (*cp)  ||  *cp == '_')
 					    {
 						    vStringPut (name, *cp);
 						    ++cp;
@@ -180,9 +180,9 @@ static void findAspTags (void)
 					else if (strncasecmp ((const char*) cp, "sub", (size_t) 3) == 0)
 					{
 						cp+=3;
-					    while (isspace ((int) *cp))
+					    while (isspace (*cp))
 						    ++cp;
-					    while (isalnum ((int) *cp)  ||  *cp == '_')
+					    while (isalnum (*cp)  ||  *cp == '_')
 					    {
 						    vStringPut (name, *cp);
 						    ++cp;
@@ -191,7 +191,7 @@ static void findAspTags (void)
 					    vStringClear (name);
 					}
 					else {
-					    while (isalnum ((int) *cp)  ||  *cp == '_')
+					    while (isalnum (*cp)  ||  *cp == '_')
 					    {
 						    vStringPut (name, *cp);
 						    ++cp;
@@ -207,11 +207,11 @@ static void findAspTags (void)
 			{
 				cp += 8;
 
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
-					while (isalnum ((int) *cp)  ||  *cp == '_')
+					while (isalnum (*cp)  ||  *cp == '_')
 					{
 						vStringPut (name, *cp);
 						++cp;
@@ -225,11 +225,11 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "sub", (size_t) 3) == 0)
 			{
 				cp += 3;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
-					while (isalnum ((int) *cp)  ||  *cp == '_')
+					while (isalnum (*cp)  ||  *cp == '_')
 					{
 						vStringPut (name, *cp);
 						++cp;
@@ -243,11 +243,11 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "dim", (size_t) 3) == 0)
 			{
 				cp += 3;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
-					while (isalnum ((int) *cp)  ||  *cp == '_')
+					while (isalnum (*cp)  ||  *cp == '_')
 					{
 						vStringPut (name, *cp);
 						++cp;
@@ -261,11 +261,11 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "class", (size_t) 5) == 0)
 			{
 				cp += 5;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
-					while (isalnum ((int) *cp)  ||  *cp == '_')
+					while (isalnum (*cp)  ||  *cp == '_')
 					{
 						vStringPut (name, *cp);
 						++cp;
@@ -279,11 +279,11 @@ static void findAspTags (void)
 			else if (strncasecmp ((const char*) cp, "const", (size_t) 5) == 0)
 			{
 				cp += 5;
-				if (isspace ((int) *cp))
+				if (isspace (*cp))
 				{
-					while (isspace ((int) *cp))
+					while (isspace (*cp))
 						++cp;
-					while (isalnum ((int) *cp)  ||  *cp == '_')
+					while (isalnum (*cp)  ||  *cp == '_')
 					{
 						vStringPut (name, *cp);
 						++cp;
